requirements.txt
Flask~=3.1.0
numpy~=2.0.2
matplotlib~=3.9.3
tensorflow~=2.18.0
pillow~=11.0.0
gunicorn
python-dotenv==0.21.0
requests~=2.32.3